+++ changelog.md
@@ -5,7 +5,7 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 ## To-Do
 
 -----------------------------------
-## Unreleased v0.2.2(2023.06.##) @master
+## Unreleased v0.3.1(2023.06.##) @master
 ### New
 ### Changed
 ### Removed
@@ -14,6 +14,11 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 ### Known Issues
 ### Notes
 ### Reference
+
+-----------------------------------
+## BT-Gyro v0.3.0(2023.06.##) @master
+### New
+- create Hm11_Init()
 
 -----------------------------------
 ## BT-Gyro v0.2.1(2023.06.26) @master